requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
praw>=7.7.0
feedparser>=6.0.0

# Data handling
python-dateutil>=2.8.0
pytz>=2023.3

# Async support
aiohttp>=3.8.0

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
coverage>=7.3.0

# Monitoring and logging
structlog>=23.1.0

# Configuration
python-dotenv>=1.0.0